+++ projects/demos/src/main/java/libshapedraw/demos/mod_LSDDemoTridentTimeline.java
@@ -12,8 +12,6 @@ import libshapedraw.shape.XrayShape;
 import libshapedraw.transform.ShapeRotate;
 import libshapedraw.transform.ShapeTranslate;
 
-import org.lwjgl.opengl.GL11;
-
 /**
  * The built-in animateStart/animateStop methods simplify some of the complex
  * things you can do with Trident Timelines. You are not forced to use those
@@ -45,8 +43,7 @@ public class mod_LSDDemoTridentTimeline extends BaseMod {
                 Color.HOT_PINK.copy(),
                 Color.HOT_PINK.copy().scaleAlpha(XrayShape.SECONDARY_ALPHA),
                 2.5F, Float.MIN_NORMAL, 5.0F);
-        GL11.glLineWidth(8.0F);
-        shape.setSlices(12).setStacks(8).setWireframe(true);
+        shape.setSlices(12).setStacks(8).setWireframe(true, 3.0F);
         Vector3 vector = new Vector3();
         shape.addTransform(new ShapeTranslate(vector));
         shape.addTransform(new ShapeRotate(90.0, Axis.X)); // upright